requirements.txt
# Core Dependencies
langchain==0.1.0
langchain-openai==0.0.5
langgraph==0.0.20
openai==1.10.0

# Gmail API
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.115.0

# Database
sqlalchemy==2.0.25

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# CLI
click==8.1.7
rich==13.7.0

# Optional: UI
streamlit==1.30.0

# Optional: Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Optional: Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0
